Extract MainState
